+++ alembic/env.py
@@ -2,7 +2,7 @@ import asyncio
 from logging.config import fileConfig
 
 from app.core.config import settings
-from app.db.schemas import metadata
+from app.db.base_class import Base
 from sqlalchemy import engine_from_config, pool
 from sqlalchemy.ext.asyncio import AsyncEngine
 
@@ -19,7 +19,7 @@ fileConfig(config.config_file_name)
 # add your model's MetaData object here
 # for 'autogenerate' support
 # from myapp import mymodel
-target_metadata = metadata
+target_metadata = Base.metadata
 
 # other values from the config, defined by the needs of env.py,
 # can be acquired:
