+++ app/db/base.py
@@ -1,3 +1,4 @@
 # Import models first so that Base will have them before being imported by Alembic
 from app.db.base_class import Base  # noqa
-from app.v1.models import User, ModelItem  # noqa
+from app.schemas.model_item import ModelItem  # noqa
+from app.schemas.user import User  # noqa
